ViPeople/requirements.txt
Django==1.6
South==0.8.4
django-vanilla-views==1.0.2
wsgiref==0.1.2
